Move more states in a different module
